// Tn_test-main/src/types/intel.ts
export interface FireData {
  lat: number;
  lon: number;
  date: Date;
  intensity: number;
  confidence: 'low' | 'nominal' | 'high';
  brightness: number;
  frp?: number; // Fire Radiative Power
}